requirements.txt
python==3.11.5
biopython==1.83
datasets==2.19.1
fair-esm==2.0.0
ipyparallel==8.8.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
torch==2.6.0
torch_geometric==2.6.1
torch-cluster==1.6.3
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-spline-conv==1.2.2
tqdm==4.65.0
transformers==4.48.0
